ajustes/coluna_preco | valor_compra de integer pra decimal + add views Adiciona views pra textos iniciativa e detalhamento que estavam quebrando

+++ database/migrations/2018_08_09_135721_create_textos_cubos_table.php
@@ -20,7 +20,7 @@ class CreateTextosCubosTable extends Migration
             $table->text('resposta_pe');
             $table->text('resposta_pf');
             $table->text('path_pdf');
-            $table->integer('valor_compra');
+            $table->decimal('valor_compra', 18, 2);
             $table->timestamps();
             $table->softDeletes();
         });
